+++ backend/core/src/test/java/com/rollthedice/backend/domain/auth/AuthTest.java
@@ -2,7 +2,7 @@ package com.rollthedice.backend.domain.auth;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
-import com.rollthedice.backend.domain.member.controller.MemberController;
+import com.rollthedice.backend.domain.member.api.MemberController;
 import com.rollthedice.backend.domain.member.dto.MemberUpdateDto;
 import com.rollthedice.backend.domain.member.service.MemberService;
 import com.rollthedice.backend.global.LoginTest;
